allow decryption using a privkey only
Previously, when decrypting we derived the fingerprint of the key to use
by assuming that the pubkey is stored alongside the privkey. In some
cases, this is simply not the case, and so we now extract the pubkey from
the privkey and then get the fingerprint from there.

This has the unfortunate side effect that in those scenarios the user may
get prompted for their passphrase twice: once to extract the pubkey and
once to actually decrypt the message.