import/merge jass version 2.1 from Twitter
This includes a complete rewrite of jass(1) in Go, which allows us to
work around a bug in Os X Maverick's openssh(1) (where it forgot how to
convert ssh keys to PKCS8 format; https://www.netmeister.org/blog/ssh2pkcs8.html)
and around a peculiarity in openssl's enc(1), where it traps all signals
after reading a wrong a password, making it impossible to terminate the
process from the controlling terminal.

This also allowed the author to learn a bit more Go and a bit more about
ssh keys and RSA encryption, so we'll call that a net-win, too.

All bugs left or newly introduced are now assumed to be original to this
program.