3.0.2: better detect when input is garbled
In older versions of jass(1), we allowed the 'message' (ie the actual
encrypted content) to be named after the input file if jass(1) was invoked
with 'jass -f file'.  In that case, the first component of the output of
jass(1) would have been named 'file'.  To account for this variable name
when decrypting, jass(1) would assume that the first chunk of the input i
the data.

We then removed this behaviour and began always encoding the message a
'message'.  This now allows us to look for that string in the input and
bail out when we don't find it, instead of exploding when we attempt to
decrypt an erroneous chunk of data.