Remove support for KEYKEEPER_URL
KeyKeeper is no longer used anywhere that I know of,
nor was it ever used outside of one company, so should
be safe to remove support for to avoid accidentally
contacting services that cannot be reached anyway.

This addresses https://github.com/jschauma/jass/issues/23
and the resurrection of https://github.com/jschauma/jass/issues/15.