add support for password managers
You can now have jass(1) fetch the passphrase for the private key
from a password manager.  Supported password managers are
- macOS keychain ('-p keychain:name')
- LastPass ('-p lpass:name')
- 1Password ('-p op:name')

For the last two, you need to install the command-line client for
the given password manager.  For 1Password, it is recommended to use
the macOS Beta with support for biometrics for unlocking; for
LastPass, you need to first have logged into LastPass ('lpass login').